# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Bagel's Repair Shop API"
  description: "API documentation for Bagel's Repair Shop"
  version: "0.0.1"
host: "127.0.0.1:5000" # Update with your server's host and port. It's in development right now
schemes:
  - "http"
consumes:
  - "application/json" #The API consumes JSON requests
produces:
  - "application/json" #The API produces JSON responses
securityDefinitions: #Defines the security scheme used in the API and how to add token authentication to endpoints
  bearerAuth: #name of the security scheme
    type: apiKey #type of security scheme
    name: Authorization
    in: header #where to find the token, in this case in the header of the request
paths: #Information about your API endpoints will go here
...
  /parts:
    post: #TOKEN REQUIRED create part route
      tags:
        - Parts
      summary: "Create a new part"
      description: "Creates a new part in the system. User will be identified via mechanic JWT token."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Part details"
          required: true
          schema:
            $ref: "#/definitions/PartCreate"
          examples:
            application/json:
              part_name: "Brake Pad"
              price: 49.99
              stock: 100
      responses:
        201:
          description: "Part created successfully"
          schema:
            $ref: "#/definitions/PartResponse"
        400:
          description: "Bad Request - Invalid input data."
          schema:
            $ref: "#/definitions/400Response"
        403:
          description: "Forbidden - Missing or invalid mechanic JWT token."
          schema:
            $ref: "#/definitions/403MechanicResponse"
    get: #get all parts
      tags:
        - Parts
      summary: "Get all parts"
      description: "Retrieves a list of all parts in the system. only accessible to authenticated mechanics."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "List of parts retrieved successfully"
          schema:
              type: array
              items:
                $ref: "#/definitions/PartResponse"
          examples:
            application/json:
              - id: 1
                part_name: "Brake Pad"
                price: 49.99
                stock: 100
        403:
          description: "Forbidden - Missing or invalid mechanic JWT token."
          schema:
            $ref: "#/definitions/403MechanicResponse"
          examples:
            application/json:
              message: "Missing or invalid token"
        404:
          description: "Not Found - No parts found in the system."
          schema:
            $ref: "#/definitions/404MechanicResponse"
          examples:
            application/json:
              message: "No parts found"

    delete: #TOKEN REQUIRED
      tags:
        - Parts
      summary: "Delete part"
      description: "Deletes a part from the system. User will be identified via the JWT token."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Part ID to delete"
          required: true
          schema:
            type: object
            properties:
              part_id:
                type: integer
            required:
              - part_id
      responses:
        200:
          description: "Part deleted successfully"
          examples:
            application/json:
              message: "Part {part_id} deleted successfully"
        400:
          description: "Bad Request - Invalid input data."
          schema:
            $ref: "#/definitions/400Response"
        403:
          description: "Forbidden - Missing or invalid mechanic JWT token."
          schema:
            $ref: "#/definitions/403MechanicResponse"
        404:
          description: "Not Found - Part not found."  
          schema:
            $ref: "#/definitions/404MechanicResponse"
  /parts/{part_id}: #TOKEN REQUIRED get, update, delete part by ID
    get:
      tags:
        - Parts
      summary: "Get part by ID"
      description: "Retrieves the details of a specific part by its ID."
      security:
        - bearerAuth: []
      parameters:
        - name: "part_id"
          in: "path"
          description: "ID of the part to retrieve"
          required: true
          type: "integer"
      responses:
        200:
          description: "Part details retrieved successfully"
          schema:
            $ref: "#/definitions/PartResponse"
        403:
          description: "Forbidden - Missing or invalid JWT token."
          schema:
            $ref: "#/definitions/403MechanicResponse"
        404:
          description: "Not Found - Part not found."
          schema:
            $ref: "#/definitions/404MechanicResponse"
    put: #update part
      tags:
        - Parts
      summary: "Update part details"
      description: "Updates the details of an existing part. User will be identified via mechanic JWT token."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Updated part details"
          required: true
          schema:
            $ref: "#/definitions/PartCreate"
      responses:
        200:
          description: "Part updated successfully"
          schema:
            $ref: "#/definitions/PartResponse"
        400:
          description: "Bad Request - Invalid input data."
          schema:
            $ref: "#/definitions/400Response"
        403:
          description: "Forbidden - Missing or invalid mechanic JWT token."
          schema:
            $ref: "#/definitions/403MechanicResponse"
        404:
          description: "Not Found - Part not found."  
          schema:
            $ref: "#/definitions/404MechanicResponse"
  parts/add_stock: #TOKEN REQUIRED
    put:
      tags:
        - Parts
      summary: "Add stock to part"
      description: "Adds stock to a specific part. User will be identified via the JWT token."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Add stock to a part"
          required: true
          schema:
            type: object
            properties:
              part_id:
                type: integer
              additional_stock:
                type: integer
            required:
              - part_id
              - additional_stock
      responses:
        200:
          description: "Stock added to part successfully"
          schema:
            $ref: "#/definitions/PartResponse"
          examples:
            application/json:
              message: "Successfully added {additional_stock} to part {part_id}. New stock: {part.stock}."
              id: 1
              part_name: "Brake Pad"
              price: 49.99
              stock: 150
        400:
          description: "Bad Request - Invalid input data."
          schema:
            $ref: "#/definitions/400Response"
        403:
          description: "Forbidden - Missing or invalid mechanic JWT token."
          schema:
            $ref: "#/definitions/403MechanicResponse"
        404:
          description: "Not Found - Part not found."  
          schema:
            $ref: "#/definitions/404MechanicResponse"
definitions: ##Information about the shape of input and output json
...
  PartCreate:
    type: object
    properties:
      part_name:
        type: string
        example: "Brake Pad"
      price:
        type: number
        format: float
        example: 49.99
      stock:
        type: integer
        example: 100
    required:
      - part_name
      - price
      - stock
  PartResponse:
    type: object
    properties:
      id:
        type: integer
      part_name:
        type: string
      price:
        type: number
        format: float
      stock:
        type: integer
  400Response: #generic 400 response
    type: object
    properties:
      message:
        type: string
        example: "invalid input data or email/phone already exists"
  403MechanicResponse: #generic 403 response 
    type: object
    properties:
      message:
        type: string
        example: "Missing or invalid token"
...
  404MechanicResponse: #generic 404 response
    type: object
    properties:
      message:
        type: string
        example: "Mechanic not found"
...